requirements.txt
asn1crypto==0.24.0
astroid==2.0.4
atomicwrites==1.2.1
attrs==18.2.0
beautifulsoup4==4.6.3
certifi==2018.10.15
cffi==1.11.5
chardet==3.0.4
cloudinary==1.14.0
coreapi==2.3.3
coreschema==0.0.4
coverage==4.5.1
cryptography==2.3.1
cssselect==1.0.3
defusedxml==0.5.0
dj-database-url==0.5.0
Django==2.1.2
django-activity-stream==0.6.5
django-allauth==0.38.0
django-cors-middleware==1.2.0
django-developer-panel==0.1.2
django-extensions==2.1.3
django-heroku==0.3.1
django-jsonfield==1.0.1
django-model-utils==3.1.2
django-notifications-hq==1.5.0
django-rest-auth==0.9.3
django-rest-swagger==2.1.0
djangorestframework==3.9.0
djangorestframework-jwt==1.11.0
gunicorn==19.9.0
idna==2.7
isort==4.3.4
itypes==1.1.0
Jinja2==2.10
jsonfield==2.0.2
lazy-object-proxy==1.3.1
lxml==4.2.5
Markdown==3.0.1
markdown2==2.3.6
MarkupSafe==1.0
mccabe==0.6.1
mock==2.0.0
more-itertools==4.3.0
oauthlib==2.1.0
openapi-codec==1.3.2
pbr==5.1.0
Pillow==5.3.0
pluggy==0.8.0
psycopg2==2.7.5
psycopg2-binary==2.7.5
py==1.7.0
pycparser==2.19
PyJWT==1.6.4
pylint==2.1.1
pylint-django==2.0.2
pylint-plugin-utils==0.4
pyOpenSSL==18.0.0
pyquery==1.4.0
pytest==3.9.2
pytest-cov==2.6.0
pytest-django==3.4.3
python3-openid==3.1.0
pytz==2018.5
readtime==1.0.6
requests==2.20.0
requests-oauthlib==1.0.0
responses==0.10.2
simplejson==3.16.0
six==1.10.0
uritemplate==3.0.0
urllib3==1.24
Werkzeug==0.14.1
whitenoise==4.1
wrapt==1.10.11
